Use canonical way to find bash via /usr/bin/env
The hardocded assumption that bash is to be found in /bin/bash does
not alwyas hold true. On a Posix system, /usr/bin/env is the canonical
way how to find the binary.

This shows up on operating systems that are very careful and specific
about what executables to make available (such as NixOS).